refactor: rename base-result-image-hover component to base-result-image
EX-5970
